+++ src/pathological/attribute_specs.clj
@@ -3,17 +3,43 @@
   (:require
     [clojure.string :as string]))
 
-(defn view [attribute-spec]
+(defn- camel->kebab-keyword [value]
   (let [camel->kebab (requiring-resolve 'pathological.utils/camel->kebab)]
-    (if (string/includes? attribute-spec ":")
-      (keyword (camel->kebab (first (string/split attribute-spec #":"))))
-      :basic)))
+    (keyword (camel->kebab value))))
 
-(defn name [attribute-spec]
-  (let [camel->kebab (requiring-resolve 'pathological.utils/camel->kebab)]
-    (if (string/includes? attribute-spec ":")
-      (keyword (camel->kebab (second (string/split attribute-spec #":"))))
-      (keyword (camel->kebab attribute-spec)))))
+(defn- kebab->camel-string [value]
+  (let [kebab->camel (requiring-resolve 'pathological.utils/kebab->camel)]
+    (str (kebab->camel value))))
+
+(defn ->attribute-spec [value]
+  (if-not (map? value)
+    (let [just-one? (fn [coll] (= (count coll) 1))
+
+          components (string/split value #":")
+          [view name-spec] (if (just-one? components)
+                             [:basic (first components)]
+                             [(camel->kebab-keyword (first components))
+                              (second components)])
+          names (set (map camel->kebab-keyword (string/split name-spec #",")))
+          name (when (just-one? names) (first names))]
+      (if name
+        {:view view
+         :name name}
+        {:view  view
+         :names names}))
+    value))
+
+(defn ->attribute-spec-string [value]
+  (let [attribute-spec (->attribute-spec value)
+        names (or (:names attribute-spec) #{(:name attribute-spec)})
+
+        view-string (kebab->camel-string (:view attribute-spec))
+        name-strings (map kebab->camel-string names)]
+    (str view-string ":" (string/join "," name-strings))))
+
+(defn view [attribute-spec] (:view (->attribute-spec attribute-spec)))
+(defn name [attribute-spec] (:name (->attribute-spec attribute-spec)))
+(defn names [attribute-spec] (:names (->attribute-spec attribute-spec)))
 
 (defn view? [attribute-spec view-or-views]
   (let [actual-view (view attribute-spec)
@@ -28,20 +54,3 @@
                      (set (map keyword name-or-names))
                      #{(keyword name-or-names)})]
     (test-names actual-name)))
-
-(defn wildcard? [spec-component]
-  (= spec-component :*))
-
-(defn specific? [spec-component]
-  (not (wildcard? spec-component)))
-
-(defn selects [actual-spec]
-  (fn [[selector-spec _]]
-    (if (= selector-spec :else)
-      true
-      (let [[selector-view selector-name] selector-spec
-            [actual-view actual-name] actual-spec]
-        (or
-          (= selector-spec actual-spec)
-          (and (wildcard? selector-view) (= selector-name actual-name))
-          (and (wildcard? selector-name) (= selector-view actual-view)))))))